refactor: SonarCloud 構造改善指摘の解消
- useTmdbSearchRequest.ts の純粋ヘルパー 8 個を tmdb-search-results.ts に抽出
- isHiddenSearchResult を 3 つのヘルパーに分解して可読性向上
- backlog-item-utils.ts の型ガード 10 個を backlog-item-validators.ts に抽出
- work-repository.ts の OMDB フィールド構築ロジックを omdb-work-fields.ts に抽出

Refs #307
