+++ src/features/backlog/work-repository.ts
@@ -9,6 +9,12 @@ import {
 } from "../../lib/tmdb.ts";
 import { fetchOmdbWorkDetails } from "../../lib/omdb.ts";
 import { buildSearchText } from "./helpers.ts";
+import {
+  buildOmdbFields,
+  buildOmdbRatings,
+  type ExistingTmdbWorkRow,
+  type OmdbFields,
+} from "./omdb-work-fields.ts";
 import type { WorkType } from "./types.ts";
 import { buildTmdbWorkUpdate, calcBackgroundFitScore, type RatingInfo } from "./work-metadata.ts";
 import {
@@ -27,18 +33,10 @@ export {
   shouldRefreshTmdbWork,
 } from "./work-repository-helpers.ts";
 
-type ExistingTmdbWorkRow = {
-  id: string;
-  last_tmdb_synced_at: string | null;
-  omdb_fetched_at: string | null;
-  imdb_id: string | null;
-  genres: string[];
-};
 type ExistingManualWorkRow = {
   id: string;
 };
 type TmdbWorkDetails = Awaited<ReturnType<typeof fetchTmdbWorkDetails>>;
-type OmdbDetails = Awaited<ReturnType<typeof fetchOmdbWorkDetails>>;
 type TmdbWorkLookup = {
   tmdbMediaType: "movie" | "tv";
   tmdbId: number;
@@ -60,10 +58,6 @@ type TmdbSyncState = {
   shouldRefreshOmdb: boolean;
   shouldSyncTmdbForOmdb: boolean;
 };
-type OmdbDecision =
-  | { type: "skip" }
-  | { type: "clear"; omdbFetchedAt: string }
-  | { type: "fetch"; imdbId: string; omdbFetchedAt: string };
 
 export async function upsertTmdbWork(
   target: TmdbSelectionTarget,
@@ -285,7 +279,7 @@ async function upsertFetchedTmdbWork(
   }
 
   const details = await fetchTmdbWorkDetails(target);
-  const omdbFields = await buildOmdbFields(details, existing);
+  const omdbFields = await buildOmdbFields(details.imdbId, existing);
 
   if (existing) {
     return updateFetchedTmdbWork(existing.id, details, omdbFields, options.parentWorkId);
@@ -391,94 +385,6 @@ async function refreshExistingOmdbFields(
   } catch {
     // OMDb 更新の失敗は TMDb の early return をブロックしない
   }
-}
-
-type OmdbFields = {
-  rotten_tomatoes_score?: number | null;
-  imdb_rating?: number | null;
-  imdb_votes?: number | null;
-  metacritic_score?: number | null;
-  omdb_fetched_at?: string;
-};
-
-function buildClearedOmdbFields(omdbFetchedAt: string): OmdbFields {
-  return {
-    rotten_tomatoes_score: null,
-    imdb_rating: null,
-    imdb_votes: null,
-    metacritic_score: null,
-    omdb_fetched_at: omdbFetchedAt,
-  };
-}
-
-function toOmdbFields(omdb: OmdbDetails, omdbFetchedAt: string): OmdbFields {
-  return {
-    rotten_tomatoes_score: omdb.rottenTomatoesScore,
-    imdb_rating: omdb.imdbRating,
-    imdb_votes: omdb.imdbVotes,
-    metacritic_score: omdb.metacriticScore,
-    omdb_fetched_at: omdbFetchedAt,
-  };
-}
-
-function buildOmdbDecision(
-  existing: ExistingTmdbWorkRow | null,
-  imdbId: TmdbWorkDetails["imdbId"],
-): OmdbDecision {
-  const omdbFetchedAt = new Date().toISOString();
-
-  if (imdbId === null) {
-    return shouldKeepExistingNullOmdbState(existing)
-      ? { type: "skip" }
-      : { type: "clear", omdbFetchedAt };
-  }
-
-  if (!imdbId || shouldSkipOmdbRefresh(existing, imdbId)) {
-    return { type: "skip" };
-  }
-
-  return { type: "fetch", imdbId, omdbFetchedAt };
-}
-
-async function buildOmdbFields(
-  details: TmdbWorkDetails,
-  existing: ExistingTmdbWorkRow | null,
-): Promise<OmdbFields> {
-  const decision = buildOmdbDecision(existing, details.imdbId);
-
-  switch (decision.type) {
-    case "skip":
-      return {};
-    case "clear":
-      return buildClearedOmdbFields(decision.omdbFetchedAt);
-    case "fetch":
-      try {
-        const omdb = await fetchOmdbWorkDetails(decision.imdbId);
-        return toOmdbFields(omdb, decision.omdbFetchedAt);
-      } catch {
-        return {};
-      }
-  }
-}
-
-function shouldKeepExistingNullOmdbState(existing: ExistingTmdbWorkRow | null) {
-  return Boolean(
-    existing && !shouldRefreshOmdbWork(existing.omdb_fetched_at) && existing.imdb_id === null,
-  );
-}
-
-function shouldSkipOmdbRefresh(existing: ExistingTmdbWorkRow | null, imdbId: string) {
-  return Boolean(
-    existing && !shouldRefreshOmdbWork(existing.omdb_fetched_at) && existing.imdb_id === imdbId,
-  );
-}
-
-function buildOmdbRatings(omdbFields: OmdbFields): RatingInfo {
-  return {
-    imdbRating: "imdb_rating" in omdbFields ? (omdbFields.imdb_rating ?? null) : null,
-    rottenTomatoesScore:
-      "rotten_tomatoes_score" in omdbFields ? (omdbFields.rotten_tomatoes_score ?? null) : null,
-  };
 }
 
 function buildTmdbUpdatePayload(
